Develop

+++ source/Auth/Operators/SessionTokenOperator.php
@@ -9,13 +9,14 @@ namespace Spiral\Auth\Providers;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use Spiral\Auth\Entities\Token;
-use Spiral\Auth\ProviderInterface;
+use Spiral\Auth\Exceptions\InvalidTokenException;
+use Spiral\Auth\TokenOperatorInterface;
+use Spiral\Auth\Providers\Session\SessionToken;
 use Spiral\Auth\TokenInterface;
 use Spiral\Auth\UserInterface;
 use Spiral\Session\SessionInterface;
 
-class SessionProvider implements ProviderInterface
+class SessionTokenOperator implements TokenOperatorInterface
 {
     /**
      * @var SessionInterface
@@ -38,6 +39,15 @@ class SessionProvider implements ProviderInterface
     }
 
     /**
+     * @param UserInterface $user
+     * @return SessionToken
+     */
+    public function createToken(UserInterface $user)
+    {
+        return new SessionToken($user->primaryKey());
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function hasToken(Request $request)
@@ -48,17 +58,21 @@ class SessionProvider implements ProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function fetchToken(Request $request, $name)
+    public function fetchToken(Request $request)
     {
-        return new Token($this->session->has($this->key), $name);
+        return new SessionToken($this->session->has($this->key));
     }
 
     /**
      * {@inheritdoc}
      */
-    public function createToken(Request $request, Response $response, UserInterface $user)
+    public function mountToken(Request $request, Response $response, TokenInterface $token)
     {
-        $this->session->set($this->key, $user->primaryKey());
+        if ($token instanceof SessionToken) {
+            throw new InvalidTokenException("Only session tokens are allowed");
+        }
+
+        $this->session->set($this->key, $token->userPK());
 
         return $response;
     }
@@ -68,6 +82,10 @@ class SessionProvider implements ProviderInterface
      */
     public function removeToken(Request $request, Response $response, TokenInterface $token)
     {
+        if ($token instanceof SessionToken) {
+            throw new InvalidTokenException("Only session tokens are allowed");
+        }
+
         $this->session->delete($this->key);
 
         return $response;
@@ -76,8 +94,12 @@ class SessionProvider implements ProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function refreshToken(Request $request, Response $response, TokenInterface $token)
+    public function updateToken(Request $request, Response $response, TokenInterface $token)
     {
+        if ($token instanceof SessionToken) {
+            throw new InvalidTokenException("Only session tokens are allowed");
+        }
+
         return $response;
     }
 }
